Rewrite disk discovery to handle EBS and NVMEs.
Amazon has just released NVMe instances, but they won't work with our
current scripts because the NVMe devices don't follow the xvdX naming
convention.

Also, we have a significant issue running instances with EBS devices
attached: the ami setup script will bundle both EBS and ephemeral
devices together in the RAID. Given that people using EBS are likely
doing so for durability reasons, this is obviously horrible. Once the
box reboots, the part of the data previously held in the ephemeral
devices will be gone, rendering the RAID useless.

To fix this, I am changing the AMI setup process to include another
helper script that will scan the instance metadata and verify which disk
belongs to each type. Based on that, it will create separate RAID arrays
grouping together disks of the same type without mixing them. NVMes are
handled correctly in that script as well.

By default, the most commong use case for EBS + ephemeral is to have
persistent data with ephemeral commitlog. We'll do that if nothing else
is communicated, but allow that to be overriden through user data.

By passing the options:

	--ebs-role=[unused/data/commitlog/all]
	--ephemeral-role=[unused/data/commitlog/all]

One will be able to change the default behavior and adapt it to one's
environment.

Fixes #20
Fixes #22

Signed-off-by: Glauber Costa <glauber@scylladb.com>
---
v3:
- merged style comments from Avi.
- Use check_call to account for errors (avi)
- retry connection a finite amount of time (avi)
v2:
- support NVMe
- error out in invalid configuration, instead of trying to continue with
  default